Added 'categories.py' for easier management of category IDs
The change to the code in the file 'categories.py' adds a dictionary of Youtube video categories and their corresponding category ids. This replaces the manual process of writing the category id, making it easier to change for people using this script as they only need to refer to the categories in the dictionary rather than memorize or look up the category id. The change in the main.py file uses the dictionary from categories.py to dynamically build the list of category_ids, depending on whether the countries are 'primary' or not. This ensures that the correct categories are fetched based on the intended target audience, without the need for manual intervention.